Make echo'd keysize match value entered.
On line 19 it says (2048) which seems to indicate that the default that will be selected is 2048. Since nothing is entered after the prompt, it is assumed that this is what the demo user used. However the next line is incongruous when it reports that the keysize is "4096 bits".

This CL fixes that.